+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/utils/LoggingCompletionCallback.kt
@@ -1,7 +1,7 @@
 package com.mapbox.search.utils
 
-import com.mapbox.search.CompletionCallback
 import com.mapbox.search.base.logger.logd
+import com.mapbox.search.common.CompletionCallback
 
 internal class LoggingCompletionCallback(private val operationName: String) : CompletionCallback<Unit> {
     override fun onComplete(result: Unit) {