+++ MapboxSearch/sdk/src/sharedTest/java/com/mapbox/search/tests_support/BlockingCompletionCallback.kt
@@ -1,6 +1,6 @@
 package com.mapbox.search.tests_support
 
-import com.mapbox.search.CompletionCallback
+import com.mapbox.search.common.CompletionCallback
 import java.util.concurrent.CountDownLatch
 
 internal class BlockingCompletionCallback<T> : CompletionCallback<T> {